Refactor logic out of big controller
